msm: kgsl: Remove active count get/put for secure context mappings
There is no long a need force the GPU on to do MMU map/unmap
for secure contexts. Voting the GPU on would force the MMU
driver to be enabled. Now that the MMU driver does its own
voting for resources, there is no need to trick the MMU on
by turning on the GPU. Remove this dead code.

Change-Id: Ib8fa38b1047ad6a12342a7f0b7bb3bd688b13c5c
Signed-off-by: Carter Cooper <ccooper@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Oktapra Amtono <oktapra.amtono@gmail.com>